consolidate shared code and use RpcInterpreter.GetSqlCommandText() to get more informative errors
